requirements.txt
Django==3.2
django-multiselectfield==0.1.12
djangorestframework==3.12.4
asgiref==3.3.4
MarkupSafe==1.1.1
pytz==2021.1
sqlparse==0.4.1
